refactor: split models into multiple files
