+++ src/lib/models/auth.ts
@@ -1,0 +1,4 @@
+/**
+ * Credentials model for Defender authentication.
+ */
+export type Credentials = { apiKey: string, apiSecret: string };
